add feature crud financial plan and category

+++ backend/src/services/transaction.service.js
@@ -2,11 +2,8 @@ const transactionModel = require('../models/transaction.model')
 const { categoryModel } = require('../models/category.model')
 const { getStartDate } = require('../utils/date')
 const UserServices = require('./user.service')
-const {walletModel} = require('../models/wallet.model')
-const {
-  BadRequestError,
-  InternalServerError,
-} = require('../core/error.response')
+const { walletModel } = require('../models/wallet.model')
+const { BadRequestError, InternalServerError, NotFoundError } = require('../core/error.response')
 const { Types } = require('mongoose')
 
 class TransactionService {
@@ -24,22 +21,20 @@ class TransactionService {
       throw new BadRequestError('Invalid wallet')
     }
     try {
-      const { transactions } = await walletModel
-        .findOne({ _id: walletId })
-        .populate({
-          path: 'transactions',
-          match: {
-            _id: !!options.offset && { $gt: options.offset },
-            createdAt: {
-              $gte: getStartDate(options.filter),
-              $lt: new Date(new Date().getTime() + 60000*420) // timezone GMT +7
-            },
+      const { transactions } = await walletModel.findOne({ _id: walletId }).populate({
+        path: 'transactions',
+        match: {
+          _id: !!options.offset && { $gt: options.offset },
+          createdAt: {
+            $gte: getStartDate(options.filter),
+            $lt: new Date(new Date().getTime() + 60000 * 420), // timezone GMT +7
           },
-          options: {
-            limit: options.limit,
-            sort: { createdAt: options.sort === 'desc' ? -1 : 1 },
-          },
-        })
+        },
+        options: {
+          limit: options.limit,
+          sort: { createdAt: options.sort === 'desc' ? -1 : 1 },
+        },
+      })
 
       console.log(transactions)
 
@@ -59,38 +54,38 @@ class TransactionService {
       const transaction = await transactionModel.findOne({ _id: transactionId })
       return transaction
     } catch (error) {
+      console.log(error)
+
       throw new InternalServerError('Get transaction error')
     }
   }
 
   static createTransaction = async ({ userId, walletId, transaction }) => {
+    const foundUser = await UserServices.findById(userId)
+    if (!foundUser || !foundUser.wallets.includes(walletId)) {
+      throw new BadRequestError('Invalid user or wallet')
+    }
+
     // check valid category
     const foundCategory = await categoryModel
-      .findOne({ _id: transaction.category._id })
+      .findOne({ 'sub_categories._id': transaction.category._id })
       .lean()
 
     if (!foundCategory) {
-      throw new BadRequestError('Invalid request')
+      throw new BadRequestError('Category not found')
     }
     const newTransaction = await transactionModel.create(transaction)
     if (!newTransaction) {
       throw new BadRequestError('Cannot create transaction')
     }
     try {
-      const foundUser = await UserServices.findById(userId)
-      if (!foundUser || !foundUser.wallets.includes(walletId)) {
-        throw new BadRequestError('Invalid user or wallet')
-      }
-
       // update wallet
       await walletModel.findOneAndUpdate(
         { _id: walletId },
         {
           $push: { transactions: newTransaction._id },
           $inc: {
-            walletBalance:
-              newTransaction.amount *
-              (newTransaction.type === 'expense' ? -1 : 1),
+            balance: newTransaction.amount * (newTransaction.type === 'expense' ? -1 : 1),
           },
         },
         { new: true }
@@ -104,41 +99,38 @@ class TransactionService {
     }
   }
 
-  static updateTransaction = async ({
-    walletId,
-    transactionId,
-    transaction,
-  }) => {
-    const wallet = await walletModel.findOne({ _id: walletId })
-    if (!wallet) {
-      throw new BadRequestError('Invalid request')
+  static updateTransaction = async ({ walletId, transactionId, update }) => {
+    const foundUser = await UserServices.findById(userId)
+    if (!foundUser || !foundUser.wallets.includes(walletId)) {
+      throw new BadRequestError('Not found wallet or user')
     }
-
+    if (update.category) {
+      const foundCategory = await categoryModel
+        .findOne({ 'sub_categories._id': update.category._id })
+        .lean()
+      if (!foundCategory) {
+        throw new BadRequestError('Category not found')
+      }
+    }
     try {
       const filter = { _id: transactionId },
-        update = transaction,
         options = { new: true, insert: true }
 
       const foundTransaction = await transactionModel.findOne(filter)
 
       const { amount, type } = foundTransaction
 
-      const updatedTransaction = await transactionModel.findOneAndUpdate(
-        filter,
-        update,
-        options
-      )
-      const newAmount =
+      const updatedTransaction = await transactionModel.findOneAndUpdate(filter, update, options)
+      const updateAmount =
         (updatedTransaction.type === 'expense'
           ? -updatedTransaction.amount
-          : updatedTransaction.amount) +
-        (foundTransaction.type === 'expense' ? amount : -amount)
+          : updatedTransaction.amount) + (foundTransaction.type === 'expense' ? amount : -amount)
       await walletModel.findOneAndUpdate(
         {
           _id: walletId,
         },
         {
-          $inc: { walletBalance: newAmount },
+          $inc: { balance: updateAmount },
         }
       )
       return updatedTransaction
@@ -149,18 +141,20 @@ class TransactionService {
   }
 
   static deleteTransactionById = async ({ walletId, transactionId }) => {
+    const foundUser = await UserServices.findById(userId)
+    if (!foundUser || !foundUser.wallets.includes(walletId)) {
+      throw new BadRequestError('Not found wallet or user')
+    }
     const wallet = await walletModel.findOne({ _id: walletId })
     if (!wallet || !wallet.transactions.includes(transactionId)) {
-      throw new BadRequestError('Invalid request')
+      throw new BadRequestError('Not found transaction')
     }
     try {
       const foundTransaction = await transactionModel.findOne({
         _id: transactionId,
       })
       const newAmount =
-        foundTransaction.type === 'expense'
-          ? foundTransaction.amount
-          : -foundTransaction.amount
+        foundTransaction.type === 'expense' ? foundTransaction.amount : -foundTransaction.amount
 
       // delete transaction
       const deletedTransaction = await transactionModel.deleteOne({
@@ -174,7 +168,7 @@ class TransactionService {
         },
         {
           $pull: { transactions: transactionId },
-          $inc: { walletBalance: newAmount },
+          $inc: { balance: newAmount },
         }
       )
       return deletedTransaction
@@ -191,6 +185,7 @@ class TransactionService {
       })
       return deletedTransactions
     } catch (error) {
+      console.log(error)
       throw new InternalServerError('Delete all transactions error')
     }
   }
